+++ packages/core/src/components/Binding/index.ts
@@ -1,0 +1,1 @@
+export * from './Binding'
