+++ packages/core/src/components/Bounds/BoundsBg.tsx
