+++ packages/core/src/components/Bounds/__tests__/corner-handle.test.tsx
@@ -1,7 +1,7 @@
 import { renderWithContext } from '~test'
 import { screen } from '@testing-library/react'
 import * as React from 'react'
-import { CornerHandle } from '../corner-handle'
+import { CornerHandle } from '../CornerHandle'
 import { TLBoundsCorner } from '~types'
 
 jest.spyOn(console, 'error').mockImplementation(() => void null)
