+++ packages/core/src/components/Bounds/__tests__/edge-handle.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { renderWithContext } from '~test'
 import { screen } from '@testing-library/react'
-import { EdgeHandle } from '../edge-handle'
+import { EdgeHandle } from '../EdgeHandle'
 import { TLBoundsEdge } from '~types'
 
 jest.spyOn(console, 'error').mockImplementation(() => void null)
