+++ packages/core/src/components/Bounds/__tests__/rotate-handle.test.tsx
@@ -1,7 +1,7 @@
 import { renderWithContext } from '~test'
 import { screen } from '@testing-library/react'
 import * as React from 'react'
-import { RotateHandle } from '../rotate-handle'
+import { RotateHandle } from '../RotateHandle'
 
 jest.spyOn(console, 'error').mockImplementation(() => void null)
