--- packages/core/src/components/bounds/index.ts
@@ -1,1 +1,0 @@
-export * from './bounds'
