+++ packages/core/src/components/page/page.tsx
@@ -3,8 +3,8 @@ import { observer } from 'mobx-react-lite'
 import * as React from 'react'
 import type { TLBinding, TLPage, TLPageState, TLShape } from '~types'
 import { useSelection, useShapeTree, useTLContext } from '~hooks'
-import { Bounds } from '~components/bounds'
-import { BoundsBg } from '~components/bounds/bounds-bg'
+import { Bounds } from '~components/Bounds'
+import { BoundsBg } from '~components/Bounds/BoundsBg'
 import { Handles } from '~components/handles'
 import { ShapeNode } from '~components/shape'
 import { ShapeIndicator } from '~components/shape-indicator'
